apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "llama-index-core>=0.12",
    "llama-index-llms-anthropic>=0.6",
    "llama-index-llms-openai>=0.3",
    "llama-index-llms-ollama>=0.5",
    "sse-starlette>=2",
    "pydantic-settings>=2",
    "python-dotenv>=1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
